Add functionality to King and Queen. Setup piece arrangements. Add testing board. Run checkForKingInRange() after piece moved.
King and Queen were not programmed, thus did not move properly.
The board did not have the correct piece arrangement due to testing,
thus an additional class made similar to the overall board but, is
optimized for testing piece behavior. The original board now has
the correct piece placement. After a piece is moved the program checks
the board for a king of opposite alliance in range of piece, then runs
gameOver().