Change process of checkmating by detecting location of king Add functionality to Pawn piece fix bug for audio fix bug for audio ON/OFF button Add some utilities
Finishing touches were added so this is a final product. Although, this is not a complete chess game, since features like castling were not added. This product can be improved in a future time.

Need bug fixes for piece graveyard - Graveyard adds piece to graveyard even if piece is not killed.